+++ R/fig_layer.R
@@ -17,6 +17,87 @@ add_layer <- function(fig, spec, dat, lname, lgroup) {
   glyph <- spec$glyph
   glyph <- underscore2camel(glyph)
   spec$glyph <- NULL
+
+  glyph_attrs <- get_glyph_attrs(spec, dat)
+
+  if(glyph == "Image") {
+    c_id <- gen_id(fig, "ColorMapper")
+    cmap <- color_mapper_model(c_id, palette = dat$palette)
+    glyph_attrs$color_mapper <- cmap$ref
+    fig$x$spec$model[[c_id]] <- cmap$model
+  }
+
+  gl_id <- gen_id(fig, c(glyph, lgroup, lname))
+  glyph_mod <- glyph_model(gl_id, glyph, glyph_attrs)
+
+  # nonselection glyph
+  ns_glyph_attrs <- glyph_attrs
+  color_ind <- which(grepl("_color", names(ns_glyph_attrs)))
+
+  for(ii in color_ind) {
+    if(!is.null(ns_glyph_attrs[[ii]])) {
+      names(ns_glyph_attrs[[ii]])[names(ns_glyph_attrs[[ii]]) == "field"] <- "value"
+      if(!is.null(ns_glyph_attrs[[ii]]$value))
+        if(!is.na(ns_glyph_attrs[[ii]]$value))
+          ns_glyph_attrs[[ii]]$value <- "#e1e1e1"
+    }
+  }
+  nsgl_id <- gen_id(fig, c("ns", glyph, lgroup, lname))
+  ns_glyph_mod <- glyph_model(nsgl_id, glyph, ns_glyph_attrs)
+
+  # hover glyph
+  hov_glyph_attrs <- glyph_attrs
+  alpha_ind <- which(grepl("_alpha", names(hov_glyph_attrs)))
+
+  for(ii in alpha_ind) {
+    if(!is.null(hov_glyph_attrs[[ii]])) {
+      names(hov_glyph_attrs[[ii]])[names(hov_glyph_attrs[[ii]]) == "field"] <- "value"
+      if(!is.null(hov_glyph_attrs[[ii]]$value))
+        if(!is.na(hov_glyph_attrs[[ii]]$value))
+          hov_glyph_attrs[[ii]]$value <- 1
+    }
+  }
+  hovgl_id <- gen_id(fig, c("hov", glyph, lgroup, lname))
+  hov_glyph_mod <- glyph_model(hovgl_id, glyph, hov_glyph_attrs)
+
+  d_id <- gen_id(fig, digest(dat))
+  dat_mod <- data_model(dat, d_id)
+
+  glr_id <- gen_id(fig, c("glyph_renderer", lgroup, lname))
+  glyph_rend_mod <- glyph_renderer_model(glr_id, dat_mod$ref, glyph_mod$ref,
+    ns_glyph_mod$ref, hov_glyph_mod$ref)
+
+  fig$x$spec$model$plot$attributes$renderers[[glr_id]] <- glyph_rend_mod$ref
+
+  fig$x$spec$model[[d_id]] <- dat_mod$model
+  fig$x$spec$model[[gl_id]] <- glyph_mod$model
+  fig$x$spec$model[[nsgl_id]] <- ns_glyph_mod$model
+  fig$x$spec$model[[hovgl_id]] <- hov_glyph_mod$model
+  fig$x$spec$model[[glr_id]] <- glyph_rend_mod$model
+
+  # add in mappings from lname to refs for use in custom js callbacks
+  # but ignore legend glyphs
+  if(!grepl("^__legend", lgroup) && length(lname) == 1) {
+    if(is.null(fig$x$spec$callback$layers))
+      fig$x$spec$callback$layers <- list()
+
+    cb <- list(
+      glyph_mod$ref,
+      ns_glyph_mod$ref,
+      hov_glyph_mod$ref,
+      dat_mod$ref,
+      glyph_rend_mod$ref
+    )
+    names(cb) <- paste(lname, c("glyph", "ns_glyph", "hov_glyph",
+      "data", "glyph_rend"), sep = "_")
+
+    fig$x$spec$callback$layers[[lname]] <- cb
+  }
+
+  fig
+}
+
+get_glyph_attrs <- function(spec, dat = NULL) {
   spec_names <- names(spec)
   spec_in_data <- spec_names %in% names(dat)
 
@@ -39,8 +120,8 @@ add_layer <- function(fig, spec, dat, lname, lgroup) {
       return(NULL)
     }
   })
-
   names(glyph_attrs) <- names(spec)
+
   if(!is.null(glyph_attrs$size))
     glyph_attrs$size$units <- "screen"
 
@@ -56,45 +137,10 @@ add_layer <- function(fig, spec, dat, lname, lgroup) {
   if(!is.null(glyph_attrs$text))
     glyph_attrs$text$field <- glyph_attrs$text$field$field
 
-  if(glyph == "Image") {
-    c_id <- gen_id(fig, "ColorMapper")
-    cmap <- color_mapper_model(c_id, palette = dat$palette)
-    glyph_attrs$color_mapper <- cmap$ref
-    fig$x$spec$model[[c_id]] <- cmap$model
-  }
+  if(!is.null(glyph_attrs$visible))
+    glyph_attrs$visible <- glyph_attrs$visible$value
 
-  gl_id <- gen_id(fig, c(glyph, lgroup, lname))
-  glyph_obj <- glyph_model(gl_id, glyph, glyph_attrs)
-
-  # nonselection glyph
-  ns_glyph_attrs <- glyph_attrs
-  color_ind <- which(grepl("_color", names(ns_glyph_attrs)))
-
-  for(ii in color_ind) {
-    if(!is.null(ns_glyph_attrs[[ii]])) {
-      names(ns_glyph_attrs[[ii]])[names(ns_glyph_attrs[[ii]]) == "field"] <- "value"
-      if(!is.null(ns_glyph_attrs[[ii]]$value))
-        if(!is.na(ns_glyph_attrs[[ii]]$value))
-          ns_glyph_attrs[[ii]]$value <- "#e1e1e1"
-    }
-  }
-  nsgl_id <- gen_id(fig, c("ns", glyph, lgroup, lname))
-  ns_glyph_obj <- glyph_model(nsgl_id, glyph, ns_glyph_attrs)
-
-  d_id <- gen_id(fig, digest(dat))
-  dat_mod <- data_model(dat, d_id)
-
-  glr_id <- gen_id(fig, c("glyph_renderer", lgroup, lname))
-  glyph_rend <- glyph_renderer_model(glr_id, dat_mod$ref, glyph_obj$ref, ns_glyph_obj$ref)
-
-  fig$x$spec$model$plot$attributes$renderers[[glr_id]] <- glyph_rend$ref
-
-  fig$x$spec$model[[d_id]] <- dat_mod$model
-  fig$x$spec$model[[gl_id]] <- glyph_obj$model
-  fig$x$spec$model[[nsgl_id]] <- ns_glyph_obj$model
-  fig$x$spec$model[[glr_id]] <- glyph_rend$model
-
-  fig
+  glyph_attrs
 }
 
 data_model <- function(dd, id = NULL) {
@@ -116,10 +162,14 @@ glyph_model <- function(id, glyph = "Circle", attrs) {
   res
 }
 
-glyph_renderer_model <- function(id, data_ref, glyph_ref, ns_glyph_ref) {
+glyph_renderer_model <- function(id, data_ref, glyph_ref,
+  ns_glyph_ref, hov_glyph_ref) {
+
   res <- base_model_object("GlyphRenderer", id)
   res$model$attributes["selection_glyph"] <- list(NULL)
   res$model$attributes$nonselection_glyph <- ns_glyph_ref
+  res$model$attributes["hover_glyph"] <- list(NULL)
+  res$model$attributes$hover_glyph <- hov_glyph_ref
   res$model$attributes["server_data_source"] <- list(NULL)
   res$model$attributes["name"] <- list(NULL)
   res$model$attributes$data_source <- data_ref