+++ R/fig_tool.R
@@ -28,18 +28,6 @@ tool_wheel_zoom <- function(fig, dimensions = c("width", "height")) {
   update_tool(fig, which = "wheel_zoom", args = list(dimensions = dimensions, plot_ref = fig$x$spec$ref))
 }
 
-#' Add "box_zoom" tool to a Bokeh figure
-#' @template tools
-#' @examples
-#' \donttest{
-#' figure() %>% ly_points(1:10) %>%
-#'  tool_box_zoom()
-#' }
-#' @export
-tool_box_zoom <- function(fig) {
-  update_tool(fig, which = "box_zoom", args = list(plot_ref = fig$x$spec$ref))
-}
-
 #' Add "save" tool to a Bokeh figure
 #' @template tools
 #' @examples
@@ -88,32 +76,141 @@ tool_reset <- function(fig) {
   update_tool(fig, which = "reset", args = list(plot_ref = fig$x$spec$ref))
 }
 
-#' Add "box_select" tool to a Bokeh figure
+#' Add "hover" tool to a Bokeh figure
 #' @template tools
-#' @param select_every_mousemove logical - should the tool's callback be triggered on every mouse move?
-#' @examples
-#' \donttest{
-#' figure() %>% ly_points(1:10) %>%
-#'  tool_box_select()
-#' }
+#' @template callback
+#' @example man-roxygen/ex-hover-custom-callback.R
 #' @export
-tool_box_select <- function(fig, select_every_mousemove = TRUE) {
-  update_tool(fig, which = "box_select", args = list(plot_ref = fig$x$spec$ref,
-    select_every_mousemove = select_every_mousemove))
+tool_hover <- function(fig, callback, ref_layer) {
+  fig %>% add_hover_callback(callback, ref_layer)
+}
+
+#' Add "tap" tool to a Bokeh figure
+#' @template tools
+#' @template callback
+#' @example man-roxygen/ex-tap-debug-callback.R
+#' @export
+tool_tap <- function(fig, callback, ref_layer) {
+  fig %>% add_tap_callback(callback, ref_layer)
 }
 
 #' Add "lasso_select" tool to a Bokeh figure
 #' @template tools
-#' @param select_every_mousemove logical - should the tool's callback be triggered on every mouse move?
+#' @param line_color,line_alpha,fill_color,fill_alpha,line_width,line_dash,level parameters to control the look of the selection bounding region
 #' @examples
 #' \donttest{
 #' figure() %>% ly_points(1:10) %>%
 #'  tool_lasso_select()
 #' }
 #' @export
-tool_lasso_select <- function(fig, select_every_mousemove = TRUE) {
-  update_tool(fig, which = "lasso_select", args = list(plot_ref = fig$x$spec$ref,
-    select_every_mousemove = select_every_mousemove))
+tool_lasso_select <- function(fig, line_color = "black", line_alpha = 1,
+  fill_color = "lightgrey", fill_alpha = 0.5,
+  line_width = 2, line_dash = c(4, 4),
+  level = "overlay") {
+
+  args <- list(line_color = line_color, line_alpha = line_alpha,
+    fill_color = fill_color, fill_alpha = fill_alpha,
+    line_dash = line_dash, line_width = line_width)
+
+  args <- get_glyph_attrs(args)
+
+  args$level <- level
+  args$xs_units <- "screen"
+  args$ys_units <- "screen"
+
+  aid <- gen_id(fig, "LassoPolyAnnotation")
+  pa_model <- base_model_object("PolyAnnotation", id = aid)
+
+  pa_model$model$attributes <- c(pa_model$model$attributes, args)
+
+  fig$x$spec$model$plot$attributes$renderers[[pa_model$ref$id]] <- pa_model$ref
+  fig$x$spec$model[[pa_model$ref$id]] <- pa_model$model
+
+  update_tool(fig, which = "lasso_select",
+    args = list(plot_ref = fig$x$spec$ref,
+    overlay = pa_model$ref))
+}
+
+#' Add "box_select" tool to a Bokeh figure
+#' @template tools
+#' @param line_color,line_alpha,fill_color,fill_alpha,line_width,line_dash,level parameters to control the look of the selection bounding region
+#' @examples
+#' \donttest{
+#' figure() %>% ly_points(1:10) %>%
+#'  tool_box_select()
+#' }
+#' @export
+tool_box_select <- function(fig,
+  line_color = "black", line_alpha = 1,
+  fill_color = "lightgrey", fill_alpha = 0.5,
+  line_width = 2, line_dash = c(4, 4),
+  level = "overlay") {
+
+  args <- list(line_color = line_color, line_alpha = line_alpha,
+    fill_color = fill_color, fill_alpha = fill_alpha,
+    line_dash = line_dash, line_width = line_width)
+
+  args <- get_glyph_attrs(args)
+
+  args$level <- level
+  args$top_units <- "screen"
+  args$bottom_units <- "screen"
+  args$left_units <- "screen"
+  args$right_units <- "screen"
+  args$render_mode <- "css"
+
+  aid <- gen_id(fig, "BoxSelectBoxAnnotation")
+  ba_model <- base_model_object("BoxAnnotation", id = aid)
+
+  ba_model$model$attributes <- c(ba_model$model$attributes, args)
+
+  fig$x$spec$model$plot$attributes$renderers[[ba_model$ref$id]] <- ba_model$ref
+  fig$x$spec$model[[ba_model$ref$id]] <- ba_model$model
+
+  update_tool(fig, which = "box_select",
+    args = list(plot_ref = fig$x$spec$ref,
+    overlay = ba_model$ref))
+}
+
+#' Add "box_zoom" tool to a Bokeh figure
+#' @template tools
+#' @param line_color,line_alpha,fill_color,fill_alpha,line_width,line_dash,level parameters to control the look of the selection bounding region
+#' @examples
+#' \donttest{
+#' figure() %>% ly_points(1:10) %>%
+#'  tool_box_zoom()
+#' }
+#' @export
+tool_box_zoom <- function(fig,
+  line_color = "black", line_alpha = 1,
+  fill_color = "lightgrey", fill_alpha = 0.5,
+  line_width = 2, line_dash = c(4, 4),
+  level = "overlay") {
+
+  args <- list(line_color = line_color, line_alpha = line_alpha,
+    fill_color = fill_color, fill_alpha = fill_alpha,
+    line_dash = line_dash, line_width = line_width)
+
+  args <- get_glyph_attrs(args)
+
+  args$level <- level
+  args$top_units <- "screen"
+  args$bottom_units <- "screen"
+  args$left_units <- "screen"
+  args$right_units <- "screen"
+  args$render_mode <- "css"
+
+  aid <- gen_id(fig, "BoxZoomBoxAnnotation")
+  ba_model <- base_model_object("BoxAnnotation", id = aid)
+
+  ba_model$model$attributes <- c(ba_model$model$attributes, args)
+
+  fig$x$spec$model$plot$attributes$renderers[[ba_model$ref$id]] <- ba_model$ref
+  fig$x$spec$model[[ba_model$ref$id]] <- ba_model$model
+
+  update_tool(fig, which = "box_zoom",
+    args = list(plot_ref = fig$x$spec$ref,
+    overlay = ba_model$ref))
 }
 
 tool_help <- function(fig, redirect = "http://hafen.github.io/rbokeh",
@@ -149,7 +246,8 @@ tool_model <- function(id, tool_name, plot_ref, ...) {
   dots <- list(...)
   dotnms <- names(dots)
   for(nm in dotnms) {
-    res$model$attributes[[nm]] <- I(dots[[nm]])
+    trns <- ifelse(is.logical(dots[[nm]]), identity, I)
+    res$model$attributes[[nm]] <- trns(dots[[nm]])
   }
 
   res