+++ R/layer_glyphs.R
@@ -23,7 +23,8 @@ ly_annular_wedge <- function(
   inner_radius = 0.1, outer_radius = 0.3,
   start_angle = 0, end_angle = 2*pi, direction = "anticlock",
   color = NULL, alpha = 1,
-  hover = NULL, url = NULL, legend = NULL, lname = NULL, lgroup = NULL, ...
+  hover = NULL, url = NULL, legend = NULL,
+  lname = NULL, lgroup = NULL, visible = TRUE, ...
 ) {
 
   validate_fig(fig, "ly_annular_wedge")
@@ -44,6 +45,7 @@ ly_annular_wedge <- function(
       legend,
       lname,
       lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -61,7 +63,8 @@ ly_annular_wedge <- function(
 
   mc <- lapply(match.call(), deparse)
 
-  make_glyph(fig, type = "annular_wedge", lname = args$info$lname, lgroup = args$info$lgroup,
+  make_glyph(fig, type = "annular_wedge", lname = args$info$lname,
+    lgroup = args$info$lgroup,
     data = args$data, data_sig = ifelse(is.null(data), NA, digest(data)),
     args = args$params, axis_type_range = axis_type_range,
     hover = args$info$hover, url = args$info$url, legend = args$info$legend,
@@ -89,7 +92,8 @@ ly_annulus <- function(
   inner_radius = 0.1, outer_radius = 0.2,
   color = NULL, alpha = 1,
   hover = NULL, url = NULL, legend = NULL,
-  lname = NULL, lgroup = NULL, ...
+  lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_annulus")
@@ -107,6 +111,7 @@ ly_annulus <- function(
       legend,
       lname,
       lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -129,7 +134,8 @@ ly_annulus <- function(
     data = args$data, data_sig = ifelse(is.null(data), NA, digest(data)),
     args = args$params, axis_type_range = axis_type_range,
     hover = args$info$hover, url = args$info$url, legend = args$info$legend,
-    xname = args$info$x_name, yname = args$info$y_name, ly_call = mc)
+    xname = args$info$x_name, yname = args$info$y_name,
+    ly_call = mc)
 }
 
 #' Add an "arc" layer to a Bokeh figure
@@ -153,7 +159,8 @@ ly_arc <- function(
   color = NULL, alpha = 1, width = 2, type = 1,
   radius = 0.2,
   start_angle = 0, end_angle = 2*pi, direction = "anticlock",
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_arc")
@@ -170,11 +177,12 @@ ly_arc <- function(
       start_angle,
       end_angle,
       direction,
-      # hover, # no hover?
-      # url, # no url?
+      # hover,
+      # url,
       legend,
       lname,
       lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -192,9 +200,9 @@ ly_arc <- function(
     fig, type = "arc", lname = args$info$lname, lgroup = args$info$lgroup,
     data = args$data, data_sig = ifelse(is.null(data), NA, digest(data)),
     args = args$params, axis_type_range = axis_type_range,
-    legend = args$info$legend, xname = args$info$x_name, yname = args$info$y_name,
-    ly_call = mc
-  )
+    hover = args$info$hover, url = args$info$url,
+    legend = args$info$legend, xname = args$info$x_name,
+    yname = args$info$y_name, ly_call = mc)
 }
 
 #' Add a "wedge" layer to a Bokeh figure
@@ -219,7 +227,9 @@ ly_wedge <- function(
   fig, x, y = NULL, data = figure_data(fig),
   radius = 0.3, start_angle = 0, end_angle = 2*pi, direction = "anticlock",
   color = NULL, alpha = 1,
-  hover = NULL, url = NULL, legend = NULL, lname = NULL, lgroup = NULL, ...
+  hover = NULL, url = NULL, legend = NULL,
+  lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_wedge")
@@ -239,6 +249,7 @@ ly_wedge <- function(
       legend,
       lname,
       lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )