+++ R/layer_lines.R
@@ -15,7 +15,8 @@
 ly_lines <- function(
   fig, x, y = NULL, data = figure_data(fig), group = NULL,
   color = "black", type = 1, width = 1, alpha = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_lines")
@@ -31,7 +32,10 @@ ly_lines <- function(
       type,
       width,
       alpha,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -121,7 +125,7 @@ ly_lines <- function(
 #' @export
 ly_segments <- function(fig, x0, y0, x1, y1, data = figure_data(fig),
   color = "black", alpha = 1, width = 1, type = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...) {
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE, ...) {
 
   validate_fig(fig, "ly_segments")
 
@@ -135,7 +139,10 @@ ly_segments <- function(fig, x0, y0, x1, y1, data = figure_data(fig),
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -184,7 +191,8 @@ ly_segments <- function(fig, x0, y0, x1, y1, data = figure_data(fig),
 ly_abline <- function(
   fig, a = NULL, b = NULL, v = NULL, h = NULL, coef = NULL,
   color = "black", alpha = NULL, width = 1, type = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...) {
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...) {
 
   validate_fig(fig, "ly_abline")
 
@@ -194,7 +202,10 @@ ly_abline <- function(
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...),
       null_data = TRUE
     )
@@ -329,7 +340,8 @@ ly_abline <- function(
 ly_curve <- function(
   fig, expr, from = NULL, to = NULL, n = 101,
   color = "black", alpha = 1, width = 1, type = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_curve")
@@ -358,7 +370,10 @@ ly_curve <- function(
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...),
       null_data = TRUE
     )
@@ -403,7 +418,8 @@ ly_contour <- function(
   x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out = ncol(z)),
   nlevels = 10, levels = pretty(range(z, na.rm = TRUE), nlevels),
   color = "black", alpha = 1, width = 1, type = 1,
-  lname = NULL, lgroup = NULL, ...
+  lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_contour")
@@ -415,7 +431,9 @@ ly_contour <- function(
       alpha,
       width,
       type,
-      lname, lgroup,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...),
       null_data = TRUE
     )
@@ -463,7 +481,8 @@ ly_ray <- function(
   fig, x, y = NULL, data = figure_data(fig),
   length = NULL, angle = 0,
   color = "black", type = 1, width = 1, alpha = NULL,
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_ray")
@@ -477,7 +496,10 @@ ly_ray <- function(
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -530,7 +552,8 @@ ly_bezier <- function(
   x0, y0, x1, y1, cx0, cy0, cx1, cy1,
   data = figure_data(fig),
   color = "black", alpha = 1, width = 1, type = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_bezier")
@@ -542,7 +565,10 @@ ly_bezier <- function(
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -597,7 +623,8 @@ ly_quadratic <- function(
   x0, y0, x1, y1, cx, cy,
   data = figure_data(fig),
   color = "black", alpha = 1, width = 1, type = 1,
-  legend = NULL, lname = NULL, lgroup = NULL, ...
+  legend = NULL, lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_quadratic")
@@ -609,7 +636,10 @@ ly_quadratic <- function(
       alpha,
       width,
       type,
-      legend, lname, lgroup,
+      legend,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )
@@ -660,7 +690,8 @@ ly_multi_line <- function(
   fig,
   xs, ys,
   color = "black", alpha = 1, width = 1, type = 1,
-  lname = NULL, lgroup = NULL, ...
+  lname = NULL, lgroup = NULL, visible = TRUE,
+  ...
 ) {
 
   validate_fig(fig, "ly_multi_line")
@@ -673,7 +704,9 @@ ly_multi_line <- function(
       width,
       type,
       # no legend?
-      lname, lgroup,
+      lname,
+      lgroup,
+      visible,
       dots = lazy_dots(...)
     )
   )