+++ man-roxygen/ex-shiny.R
@@ -2,7 +2,6 @@
 # courtesy of Joe Cheng from RStudio
 library("shiny")
 library("rbokeh")
-library("htmlwidgets")
 
 ui <- fluidPage(
   rbokehOutput("rbokeh")
@@ -14,6 +13,34 @@ server <- function(input, output, session) {
     invalidateLater(1000, session)
     figure() %>%
       ly_points(jitter(cars$speed), jitter(cars$dist))
+  })
+}
+
+shinyApp(ui, server)
+
+
+library("shiny")
+library("rbokeh")
+
+ui <- fluidPage(
+  rbokehOutput("rbokeh", width = 500, height = 540),
+  textOutput("x_range_text")
+)
+
+server <- function(input, output, session) {
+  output$rbokeh <- renderRbokeh({
+    figure() %>% ly_points(1:10) %>%
+      x_range(callback = shiny_callback("x_range"))
+  })
+
+  output$x_range_text <- reactive({
+    xrng <- input$x_range
+    if(!is.null(xrng)) {
+      paste0("factors: ", xrng$factors, ", start: ", xrng$start,
+        ", end: ", xrng$end)
+    } else {
+      "waiting for axis event..."
+    }
   })
 }